distributed into separte modules and includes command line arguments
